+++ src/language/client.ts
@@ -3,6 +3,7 @@ import * as vscode from 'vscode';
 import { LanguageClient, ServerOptions, TransportKind, LanguageClientOptions } from 'vscode-languageclient';
 import { ExtensionContext } from 'vscode';
 import { IConnection } from '../common/IConnection';
+import { EditorState } from '../common/editorState';
 
 export default class PostgreSQLLanguageClient {
 
@@ -25,6 +26,7 @@ export default class PostgreSQLLanguageClient {
     };
 
     this.client = new LanguageClient('postgres', 'PostgreSQL Service', serverOptions, clientOptions);
+    this.client.onReady().then(() => { EditorState.connection = null; });
     let disposable = this.client.start();
     context.subscriptions.push(disposable);
   }