+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to the "vscode-postgres" extension will be documented in this file.
 
 ## [Unreleased]
+### Fixed
+- Connection button, now properly asks for the connection instead of the database.
+
 ### Added
 - Syntax highlighting
 - Connection aware code completion (keywords, functions, tables, and fields)